+++ packages/common/src/types/types.ts
@@ -1,0 +1,9 @@
+export interface CompilableQuery<T> {
+  execute(): Promise<T[]>;
+  compile(): CompiledQuery;
+}
+
+export interface CompiledQuery {
+  readonly sql: string;
+  readonly parameters: ReadonlyArray<unknown>;
+}