+++ packages/react/src/hooks/usePowerSyncWatchedQuery.ts
@@ -3,6 +3,8 @@ import React from 'react';
 import { usePowerSync } from './PowerSyncContext';
 
 /**
+ * @deprecated use {@link useQuery} instead.
+ *
  * A hook to access the results of a watched query.
  * @example
  * export const Component = () => {