+++ demos/example-vite/CHANGELOG.md
@@ -1,5 +1,17 @@
 # example-vite
 
+## 0.0.17
+
+### Patch Changes
+
+- @powersync/web@0.5.2
+
+## 0.0.16
+
+### Patch Changes
+
+- 371e8ce: Updated Vite Demo apps' `include` entries to use nested dependency syntax, fixes issue with CJS nested dependencies.
+
 ## 0.0.15
 
 ### Patch Changes