+++ demos/vue-supabase-todolist/CHANGELOG.md
@@ -1,5 +1,20 @@
 # vue-supabase-todolist
 
+## 0.0.6
+
+### Patch Changes
+
+- a5550b2: Using new Vue `useQuery` and `useStatus` composables instead of deprecated ones.
+- Updated dependencies [a5550b2]
+  - @powersync/vue@0.1.0
+  - @powersync/web@0.5.2
+
+## 0.0.5
+
+### Patch Changes
+
+- 371e8ce: Updated Vite Demo apps' `include` entries to use nested dependency syntax, fixes issue with CJS nested dependencies.
+
 ## 0.0.4
 
 ### Patch Changes