+++ packages/web/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @powersync/web
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [c94be6a]
+- Updated dependencies [2f1e034]
+- Updated dependencies [21801b9]
+  - @powersync/common@1.7.0
+
 ## 0.5.1
 
 ### Patch Changes